requirements.txt
alabaster==0.7.7
astroid==1.4.2
Babel==2.1.1
bibtexparser==0.6.1
click==0.7
colorama==0.3.5
Cython==0.23.2
django-mongodb-engine==0.6.0
djangotoolbox==1.8.0
docutils==0.12
Flask==0.10.1
Flask-Admin==1.0.7
Flask-HTTPAuth==2.2.1
Flask-SQLAlchemy==1.0
gitdb==0.6.4
GitPython==1.0.1
itsdangerous==0.24
Jinja2==2.7.2
lazy-object-proxy==1.2.1
MarkupSafe==0.19
oauthlib==1.0.3
Pygments==2.0.2
PyInstaller==3.0
pylint==1.5.1
pymongo==3.0.3
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.5.0
six==1.10.0
smmap==0.9.0
snowballstemmer==1.2.0
Sphinx==1.3.3
sphinx-rtd-theme==0.1.9
SQLAlchemy==0.9.4
Werkzeug==0.9.4
wheel==0.24.0
wrapt==1.10.6
WTForms==1.0.5
yolk==0.4.3
